requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
Django==4.1.2
djangorestframework==3.14.0
executing==1.1.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
pure-eval==0.2.2
Pygments==2.13.0
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
traitlets==5.4.0
tzdata==2022.5
wcwidth==0.2.5
